update build tools and fix package names
