+++ example/build.gradle
@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.2'
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
 
     defaultConfig {
         applicationId "momanii.hani.supernova_emoji"
-        minSdkVersion 9
-        targetSdkVersion 25
+        minSdkVersion 14
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -19,6 +19,6 @@ android {
 }
 
 dependencies {
-    compile project(':supernova-emoji-library')
-    //compile 'com.android.support:appcompat-v7:25.3.1'
+    implementation project(':supernova-emoji-library')
+    implementation 'com.android.support:appcompat-v7:28.0.0'
 }
