+++ example/src/main/res/layout/activity_main.xml
@@ -29,7 +29,7 @@
         android:src="@android:drawable/ic_menu_send"
         />
 
-    <hani.momanii.supernova_emoji_library.Helper.EmojiconEditText
+    <hani.momanii.supernova_emoji_library.helper.EmojiconEditText
         android:id="@+id/emojicon_edit_text"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -41,7 +41,7 @@
         emojicon:emojiconSize="28sp"/>
 
 
-    <hani.momanii.supernova_emoji_library.Helper.EmojiconEditText
+    <hani.momanii.supernova_emoji_library.helper.EmojiconEditText
         android:id="@+id/emojicon_edit_text2"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -60,7 +60,7 @@
         android:checked="false"
         android:text="Use System Default?"/>
 
-    <hani.momanii.supernova_emoji_library.Helper.EmojiconTextView
+    <hani.momanii.supernova_emoji_library.helper.EmojiconTextView
         android:id="@+id/textView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
