+++ supernova-emoji-library/src/main/java/hani/momanii/supernova_emoji_library/helper/EmojiAdapter.java
@@ -16,7 +16,7 @@
 
 
 
-package hani.momanii.supernova_emoji_library.Helper;
+package hani.momanii.supernova_emoji_library.helper;
 
 import android.content.Context;
 import android.view.View;
