+++ supernova-emoji-library/src/main/java/hani/momanii/supernova_emoji_library/helper/EmojiconEditText.java
@@ -14,15 +14,13 @@
  * limitations under the License.
  */
 
-package hani.momanii.supernova_emoji_library.Helper;
+package hani.momanii.supernova_emoji_library.helper;
 
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.text.SpannableStringBuilder;
-import android.text.TextUtils;
 import android.text.style.DynamicDrawableSpan;
 import android.util.AttributeSet;
-import android.widget.TextView;
+import android.widget.EditText;
 
 import hani.momanii.supernova_emoji_library.R;
 
@@ -30,53 +28,41 @@ import hani.momanii.supernova_emoji_library.R;
  * @author Hieu Rocker (rockerhieu@gmail.com).
  * @author Hani Al Momani (hani.momanii@gmail.com)
  */
-public class EmojiconTextView extends TextView {
+public class EmojiconEditText extends EditText {
     private int mEmojiconSize;
     private int mEmojiconAlignment;
     private int mEmojiconTextSize;
-    private int mTextStart = 0;
-    private int mTextLength = -1;
     private boolean mUseSystemDefault = false;
 
-    public EmojiconTextView(Context context) {
+    public EmojiconEditText(Context context) {
         super(context);
-        init(null);
+        mEmojiconSize = (int) getTextSize();
+        mEmojiconTextSize = (int) getTextSize();
     }
 
-    public EmojiconTextView(Context context, AttributeSet attrs) {
+    public EmojiconEditText(Context context, AttributeSet attrs) {
         super(context, attrs);
         init(attrs);
     }
 
-    public EmojiconTextView(Context context, AttributeSet attrs, int defStyle) {
+    public EmojiconEditText(Context context, AttributeSet attrs, int defStyle) {
         super(context, attrs, defStyle);
         init(attrs);
     }
 
     private void init(AttributeSet attrs) {
+        TypedArray a = getContext().obtainStyledAttributes(attrs, R.styleable.Emojicon);
+        mEmojiconSize = (int) a.getDimension(R.styleable.Emojicon_emojiconSize, getTextSize());
+        mEmojiconAlignment = a.getInt(R.styleable.Emojicon_emojiconAlignment, DynamicDrawableSpan.ALIGN_BASELINE);
+        mUseSystemDefault = a.getBoolean(R.styleable.Emojicon_emojiconUseSystemDefault, false);
+        a.recycle();
         mEmojiconTextSize = (int) getTextSize();
-        if (attrs == null) {
-            mEmojiconSize = (int) getTextSize();
-        } else {
-            TypedArray a = getContext().obtainStyledAttributes(attrs, R.styleable.Emojicon);
-            mEmojiconSize = (int) a.getDimension(R.styleable.Emojicon_emojiconSize, getTextSize());
-            mEmojiconAlignment = a.getInt(R.styleable.Emojicon_emojiconAlignment, DynamicDrawableSpan.ALIGN_BOTTOM);
-            mTextStart = a.getInteger(R.styleable.Emojicon_emojiconTextStart, 0);
-            mTextLength = a.getInteger(R.styleable.Emojicon_emojiconTextLength, -1);
-            mUseSystemDefault = a.getBoolean(R.styleable.Emojicon_emojiconUseSystemDefault, mUseSystemDefault);
-            a.recycle();
-        }
         setText(getText());
     }
 
     @Override
-    public void setText(CharSequence text, BufferType type) {
-        if (!TextUtils.isEmpty(text)) {
-            SpannableStringBuilder builder = new SpannableStringBuilder(text);
-            EmojiconHandler.addEmojis(getContext(), builder, mEmojiconSize, mEmojiconAlignment, mEmojiconTextSize, mTextStart, mTextLength, mUseSystemDefault);
-            text = builder;
-        }
-        super.setText(text, type);
+    protected void onTextChanged(CharSequence text, int start, int lengthBefore, int lengthAfter) {
+        updateText();
     }
 
     /**
@@ -84,7 +70,12 @@ public class EmojiconTextView extends TextView {
      */
     public void setEmojiconSize(int pixels) {
         mEmojiconSize = pixels;
-        super.setText(getText());
+
+        updateText();
+    }
+
+    private void updateText() {
+        EmojiconHandler.addEmojis(getContext(), getText(), mEmojiconSize, mEmojiconAlignment, mEmojiconTextSize, mUseSystemDefault);
     }
 
     /**
