+++ supernova-emoji-library/src/main/res/layout/emojicon_item.xml
@@ -19,7 +19,7 @@
     android:layout_width="wrap_content"
     android:layout_height="wrap_content">
 
-    <hani.momanii.supernova_emoji_library.Helper.EmojiconTextView
+    <hani.momanii.supernova_emoji_library.helper.EmojiconTextView
         android:layout_gravity="center"
         android:id="@+id/emojicon_icon"
         android:layout_width="36dip"
